pyproject.toml
[tool.poetry]
name = "mambot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
wonda = "^0.5.7"
tortoise-orm = "^0.19.3"
envparse = "^0.2.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
